refactor: update the Pedersen hash function
Simplify the Pedersen hash function by introducing panics for a broken
contract and remove a check for a condition that never evaluates to
true. This not only improves the efficiency of the function but also
simplifies its signature to only return the digest instead of the digest
and an error.

This is made possible by the fact that the condition in the inner loop
of pedersen.go that checks whether the shift point is compared to itself
is never true as `pt2` will only ever the assigned to the points in the
range 2..=505.

The hexadecimal strings that represent the points are also hardcoded in
the init function in order to reduce the memory footprint of the
package along with other redundant global variables.

BREAKING CHANGE: `Digest` and `ArrayDigest` now return a single argument

Any previous calls to these functions should discard their error
handling logic.