feat: add binary trie (#129)
* feat: create binary trie

* refactor: implement delete method

* docs: update documentation

* test: add tests for exported methods

* refactor: substitute recursive put with loop

This also fixes an issue where deleting certain keys would not result in
an updated commitment value.

* test: satisfy code coverage tool

* refactor: minor changes to improve clarity

* fix: fix issue caused by overflow

* test: add test for state commitment

* test: improve test coverage

Cater for the assumption that a key-value pair where the value is zero
implies key deletion.

* fix: fix typo

* fix: fix erroneously reversed keys in trie.Put

* refactor: update the Pedersen hash function

Simplify the Pedersen hash function by introducing panics for a broken
contract and remove a check for a condition that never evaluates to
true. This not only improves the efficiency of the function but also
simplifies its signature to only return the digest instead of the digest
and an error.

This is made possible by the fact that the condition in the inner loop
of pedersen.go that checks whether the shift point is compared to itself
is never true as `pt2` will only ever the assigned to the points in the
range 2..=505.

The hexadecimal strings that represent the points are also hardcoded in
the init function in order to reduce the memory footprint of the
package along with other redundant global variables.

BREAKING CHANGE: `Digest` and `ArrayDigest` now return a single argument

Any previous calls to these functions should discard their error
handling logic.

* refactor: minor optimisations

Introduce minor performance optimisations aimed at reducing the amount
of time it takes to process a block.

This is done by preferring much more efficient alternatives to some
functions such as:

  - Comparing against the length field of a slice instead of the entire
    slice when querying the root node.
  - Replacing string formatting with appending raw bytes or using
    dedicated type formatting functions for basic types as the latter
    uses reflection and as a result is more costly.

* tests: delete Pedersen hash limit test

Remove the test that checks whether the inputs are in the required
field. This function now panics if that's the case as it is considered
to be a broken contract.

* tests: comment out array digest benchmark

Comment out the array digest benchmark which is currently failing.

* tests: satisfy coverage tool

Add a test for querying the commitment of an empty trie.

Co-authored-by: Marcos Antonio Maceo <35319980+stdevMac@users.noreply.github.com>