Fix state commitment issue on L1 sync (#208)
* Fix mismatched fact issue

If we cancel execution while processing a Starknet block that has
a corresponding LogStateUpdate in the same Ethereum block as the
previous Starknet block's LogStateUpdate (e.g. Starknet block 20)
and restart, we will get an error "state commitments do not
match". That is because we incorrectly populate `factVal` in
`getFactInfo` with information from both LogStateUpdates, rather
than the one corresponding to the current sequence number.

This fix ensures that we correctly filter the LogStateUpdates for
the one that corresponds to the current StateTransitionFact.

* Format