Add missing lru functionality, behaviour tests and refactor
A pointer to RWMutex is used so that when LRUCache is passed the mutex
doesn't cause a panic, since mutexes cannot be copied. Also, by using
RWMutex it allows read operations to be executed concurrently, while
only write operations are protected.

Move LRUCache into its own package and rename the LRUCache struct to
Cache.

Rename start and end fields of cacheNodes to front and back,
respectively.

Add behaviour tests to assert lru behaviour.

Cache of capacity less than 1 cannot be created.

If a user calls Put() with a node whose key already exists in the cache
but the value is different from the existing node in the cache then it
will be deleted and replaced with the new node. This will not change the
size of the cache.