Delete processes and services (#288)
Removing processes and services to make it easier to run and maintain juno.

Services were implemented as singletons which lead to side effects when they are
used and obscured how they were created and used.

The process struct provided a concurrency wrapper around the HTTP servers and
service calls such that each call to the functions of service would add new
process in the process handler and remove the process when the function
returned. This lead to concurrent bugs, such as panic on a negative wait group
call.