Shutdown multiple sync go-routines gracefully (#384)
In the sync package, multiple go-routines are started, however, they
were not closed properly. Hence, when `synchronizer.Close()` was called
those go-routines continued to run, which led to panic in the database
as they would try to store/retrieve information. This commit ensures the
go-routines exit before Juno shuts down.

Co-authored-by: Tshaka Eric Lekholoane <_@tshaka.co>