Add a basic GatewayClient
Currently only supports "get_state_update" endpoint because this
was started to be able to query the gateway for Trie test data.

Could be used a template for future endpoints

The change in felt.go is warranted because gateway inconsistently
omits "0x" prefix for some hex values. So if we fail to parse a hex
value as field element, we try again hoping that it is a hexadecimal
with a missing prefix