Replace felt.Felt with fp.Element
Since we are replacing our own `felt` package with the `stark-curve`
package from gnark-crypto, we can no longer deal with the feeder
gateway's inconsistent hex string representations in the source code. To
work around this, we implement a custom JSON unmarshaller using
reflection.

Otherwise, this is virtually a drop-in replacement.