Switch to using gnark-crpyto FP implementation
We had to wrap it with our own Felt type to be able to implement a
custom JSON unmarshaler. To avoid "leaking" any fp.Element typed
objects the wrapped field is private and not promoted. This forces
us to forward some functions to the underlying type. I forwarded the
smallest set of functions that we use currently.