Add db.Transaction.Seek()
This will be particularly useful with historical state storage.

Assume that we store changes on a storage address in the database
as the height of the change as the key and the old value as the value.

So our DB will look like;

Height of the change | old value
       5             |    0
      99             |    1
      101            |    2

To query the value for height 10 for example, calling Seek(10) will return
the second entry as the response. ie {99, 1}. Which represents the value
of that storage address between height 5 and 99, which includes height 10.