Only verify Cairo1 class hashes and dont ignore failures
We used to try to verify both Cairo1 and Cairo0 classes. But Cairo0
classes are really hard to verify since their hashing process is really
not well defined and easy to replicate. Since they are deprecated, no need to
bother with trying to verify them.

Cairo1 class hashes are easy to recalculate, so this commit makes it
a requirement before accepting a new block