Remove unnecessary db.ContractRootKey bucket
We can use db.ContractStorage+<address> as the root key.

With this change, we manage root keys the same for both the state trie
and the contract storage tries, enabling us to remove duplicated code
in future commits.

Warning: this is a breaking db layout change. Previous DBs are not
compatible.