Remove unnecessary db.ContractRootKey bucket
We can use db.ContractStorage+<address> as the root key.

With this change, we manage root keys the same way for both the state trie
and the contract storage tries, enabling us to remove duplicated code
in future commits.

This is a breaking db layout change. An associated migration is provided.