Set the default data directory correctly (#1249)
We now use cwd/juno as the data directory.

Previously, we set the default to the empty string "" and manually set
the default directory in `node.New`. This meant the incorrect default
was printed in two cases:

1. `--help`
2. When printing the config on boot.

Finally, this centralizes all of our config management to the
`cmd/juno` package; if `juno.NewCmd.Execute` returns without an error,
the config is structurally valid.