Update snapshot link in snapshots.md
Have the link pointing to the latest available snapshot.

In order for this to work, I had to remove the size and what's the latest block from the table, so we won't have to manually update this information everytime.

I also included a way to know how big the snapshot is via script, so this information will always be up-to-date by the time someone runs the command