fix: Remove unnecessary go in matrix
CI is getting the go version from go.mod, and there
is no reference for matrix.go.  Therefore we can
conclude that this is not being used.

The annoying part with this is when upgrading the go
version, the job will be a different name, and someone
needs to manually go to the protected rule for the main
branch and update what's the name of the tests