feat: Update transactionCommitmentPoseidon function to handle different transaction types (#2153)
* feat: Update transactionCommitmentPoseidon function to handle different transaction types

The code changes in `transactionCommitmentPoseidon` function update the logic to handle different transaction types. With the changes, the function now checks the type of the transaction and updates the digest accordingly. For `DeployTransaction` and `L1HandlerTransaction`, the digest is updated with a zero value, while for other transaction types, the digest is updated with the transaction signature.

* Added tests

* Replace dynamic hashes with hardcoded value in test

---------

Co-authored-by: Kirill <paltsev.kir@gmail.com>