Add workflow to close stale issues and PRs
Introduce a GitHub Actions workflow to automatically close stale
issues and pull requests. This helps maintain repository hygiene
by reducing clutter from inactive items. The workflow marks issues
as stale after 60 days of inactivity and closes them 14 days later
if no further activity occurs.

The issues and PRs that have a milestone assigned will not be looked
at by the bot. This means that it is still possible to plan work that
won't be worked for some time.  Also, if the tag 'no-stale' is present,
it will also be ignored.