fix(rpc): failed fee estimation due to absence of l1_data_gas (#2584)
Fixes #2577

In RPC v8, L1Gas, L2Gas and L1DataGas are part of the spec and required.
In RPC v6 and v7, only L1Gas and L2Gas are part of the spec and required.
Blockifier will throw an error if L1DataGas is absent and L2Gas is non-zero.
To avoid that, if L2Gas is non-zero, we set the L1DataGas resource to zero.