fix(cmd/juno): Fix cmd env tests (#2665)
Recently, t.Setenv() was introduced to allow temporary setting of
environment variables cmd/juno_test.go. However, it was not correctly
implemented because it would only set the environment variable which
was being tested. This meant if an environment variable has been set
on the system which is not in the list of the tests, then the test would
fail.

The solution is to unset all the variables before the test which are
automatically reset to their original value after the test.